+++ src/ck-core/register.C
@@ -60,7 +60,7 @@ int CkRegisterEp(const char *name, CkCallFnPtr call, int msgIdx, int chareIdx,
 }
 
 extern "C"
-int CkRegisterChare(const char *name, int dataSz)
+int CkRegisterChare(const char *name, size_t dataSz)
 {
   return _chareTable.add(new ChareInfo(name, dataSz));
 }
